+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Made Watchdog default logger
+- Updated Watchlog logger
+
 ## [0.1.0] - 2024-10-21
 
 - Moved drush services into own yml file (drush 12)